Correção do convertYpxToValue para contemplar double

+++ src/main/java/br/com/dbsoft/ui/component/charts/DBSCharts.java
@@ -229,12 +229,12 @@ public class DBSCharts extends DBSUIInput implements NamingContainer{
 	 * @param pPxValue
 	 * @return
 	 */
-	public Integer convertYPxToValue(Double pPxValue){
+	public Double convertYPxToValue(Double pPxValue){
 		BigDecimal xValue = DBSNumber.divide(DBSNumber.multiply(pPxValue, 
 	 														    getTotalValue()),
 											 getChartHeight());
 		xValue = DBSNumber.add(xValue, getMinValue());
-		return xValue.intValue();
+		return xValue.doubleValue();
 	}
 
 	public void setChartWidthHeight(TYPE pType){
